[MOVE] website_sale: moves the product input files
This commit changes the file structure of the product input snippet.

task-2189645
