+++ addons/website_sale/views/templates.xml
@@ -47,6 +47,7 @@
             <script type="text/javascript" src="/website_sale/static/src/js/website_sale.editor.js"></script>
             <script type="text/javascript" src="/website_sale/static/src/js/website_sale_form_editor.js"></script>
             <script type="text/javascript" src="/website_sale/static/src/js/tours/website_sale_shop_frontend.js"></script>
+            <script type="text/javascript" src="/website_sale/static/src/snippets/s_products_searchbar/options.js"></script>
         </xpath>
     </template>
